Include annotations on test methods in output (#7)
DexParser now has a new top-level entry point method
that will return not only the fully qualified test
names found in the given apk, but also the fully qualified
names of any annotations that apply to the test (either
class or method level annotations for that test).

Note that the parser api that writes output to a file
on disk has not currently been changed; it still only
writes the fully qualified test names and ignores the
annotations. Annotation retrieval is only supported
when using dex-test-parser as a Java dependency, not
when used as a command line tool. It should be easy for
someone to write a wrapper around dex-test-parser to output
the annotations via a command line tool if they need it,
and this saves dex-test-parser from needing to maintain
a more complicated serialized format of the annotation
output.

This commit has the side effect of updating to Kotlin
1.1.2 and fixing all the new compiler warnings.

Fixes #1