Add support for parsing and reading encoded values
This adds support for reading the values included in annotations, as opposed to just the name and presence. It changes the TestMethod
class to contain a new list of TestAnnotation objects, each of which represents a single annotation with its name and all values included.
To do this, we introduce the concept of Encoded and Decoded value classes. The EncodedValue class matches up with the spec, and the DecodedValue
class converts it to a more usable format.

A few of the complex encoded values (fields, methods, etc) still need to be implemented in the DecodedValue class, but I wanted to get this up
since those are relatively rarely used. Implementing these will require the addition of new data classes to represent those types, since they
are not primitives.