Add Junit tests for parser
Mostly added tests for annotation value parsing. Also setup the project so that we can write unit tests that can run against
the sample app's apk. This involved some configuration in the build.gradle to ensure that apk is up to date, and the environment
is configured the same when running from gradle or the IDE. Use the path to the apk defined in DexParserShould when loading the apk
for testing.

Also fixed an issue with parsing floating point values, which was related to how their values are padded to the right which is abnormal