Add support for parsing and reading encoded values (#11)
* Add support for parsing and reading encoded values

This adds support for reading the values included in annotations, as opposed to just the name and presence. It changes the TestMethod
class to contain a new list of TestAnnotation objects, each of which represents a single annotation with its name and all values included.
To do this, we introduce the concept of Encoded and Decoded value classes. The EncodedValue class matches up with the spec, and the DecodedValue
class converts it to a more usable format.

A few of the complex encoded values (fields, methods, etc) still need to be implemented in the DecodedValue class, but I wanted to get this up
since those are relatively rarely used. Implementing these will require the addition of new data classes to represent those types, since they
are not primitives.

* Update parser code based on review comments

* Add Junit tests for parser

Mostly added tests for annotation value parsing. Also setup the project so that we can write unit tests that can run against
the sample app's apk. This involved some configuration in the build.gradle to ensure that apk is up to date, and the environment
is configured the same when running from gradle or the IDE. Use the path to the apk defined in DexParserShould when loading the apk
for testing.

Also fixed an issue with parsing floating point values, which was related to how their values are padded to the right which is abnormal

* Fix ValidTests list for the new test method

* Fixing imports and various review comments