Add support for test classes that extend other test classes - #13 (#16)
* Add support for test classes that extend other test classes

This supports both abstract and concrete superclasses.

Normally, test methods are found via reflection which is why methods defined from the super class would be included. This
adds the same behavior to dex-test-parser to match with what is expected when running tests via adb directly.

* Fix test to include the new abstract test methods in expectation

* Fix naming, remove superfluous logs

* Bump version to 2.0

This is a breaking change to the methods to find JUnit4 methods. Also updated the changelog to indicate this

* Update changelog styling

We want to have a section for the current snapshot, so people can add changes there as they are added