Fix reporting for default interface methods and private methods
Previously, we were reporting both default interface methods and private methods that were annotated with @Test
as valid, when in reality neither is true. To fix this, we can take advantage of the differentiation the dex file
format makes in the class_data_item struct, which contains 2 sets of methods: direct_methods and virtual_methods.
For tests, all we really care about are the virtual methods, which are "none of static, private, or constructor"
methods on the class. So this updates our logic for both junit3 and junit4 to only look in the virtual_methods set
for tests.