Fix reporting for default interface methods and private methods (#22)
* Fix reporting for default interface methods and private methods

Previously, we were reporting both default interface methods and private methods that were annotated with @Test
as valid, when in reality neither is true. To fix this, we can take advantage of the differentiation the dex file
format makes in the class_data_item struct, which contains 2 sets of methods: direct_methods and virtual_methods.
For tests, all we really care about are the virtual methods, which are "none of static, private, or constructor"
methods on the class. So this updates our logic for both junit3 and junit4 to only look in the virtual_methods set
for tests.

* Add tests for classes that implement interfaces with default @Test methods

Added a new concrete class that implements the InterfaceWithDefaultMethods interface. There are 2
methods in that now, one that is overridden by the new class, and one that is not. Both should be reported.

* Bump number of expected tests in jvm tests

We added 2 new expected tests for the default interface methods support