Bump to AGP 3.2.1, add Java8 tests
We needed to bump our AGP version past 3.0 so that we could use Java8 in the sample app for tests.
There was an issue where one of our buildsrc dependencies was forcing a lower version of guava than
what AGP expected, so to fix it we force it up to version 21 to make everyon happy.