Fix #21 bufferoverflow when parsing test annotations (#24)
* Fix #21 bufferoverflow when parsing test annotations

We should never have been seeing negative values for the size of annotations, as the byte that they
are stored in is unsigned. However, in the process of converting to the byte to an int (kotlin has no
bitwise operators available to byte types), we were padding with a bunch of 1s causing it to be read
as a signed number. Forcing it down to an unsigned 8 bit value resolves this issue and lets us read the
values out correctly.

* Move the 8bit masking to Byte