Add @JvmOverloads to DexParser entry points
PR #50 tried to maintain backwards compatibility while adding support
for passing custom test annotations, and it did successfully do this for
the CLI case where the dex-test-parser jar is executed via the main
method.

Unfortunately, it was a breaking change for the library use case, where
dex-test-parser is added as a compile dependency to a JVM project and
the direct entry points on DexParser are used. This change adds a
default value for the new customAnnotations parameter along with
@JvmOverloads for Java consumers.