Use signing.required API instead of if condition
This has the advantage that if there is an error introduced in the
signing configuration, it will still be caught locally at Gradle
configuration time. With the old version of the code, Gradle syntax
errors would never be noticed in local development because code inside
the if statement would never be executed.