Differentiate overridden vs inherited super test methods (#71)
* Differentiate overridden vs inherited super test methods

There is a behavior difference between the case of a superclass test method
that is not overridden by the subclass, and one that is - when it is not overridden,
the actual super method is used, and so all of its annotations remain in tact. However
when it _is_ overridden, only inherited annotations are retained. This adds support for
handling this difference by splitting the cases when parsing.

Also updated tests to be more robust, there were some noticeable gaps.

Fixes #69