Fix BufferUnderflowException when parsing dex file (#73)
classAnnotationsOff can be 0 according to this. Missing this check causes the crash in #72.

class_annotations_off
offset from the start of the file to the annotations made directly on the class, or 0 if the class has no direct annotations. The offset, if non-zero, should be to a location in the data section. The format of the data is specified by "annotation_set_item" below.